moving unit tests next to their targets
